Move scheduler to standalone
